theme: remove redundant styling classes
* fix accordion field in deposit
* theme: refactor CSS classes
* refactor javascript elements
