+++ invenio_app_rdm/records_ui/templates/semantic-ui/invenio_app_rdm/records/dashboard.html
@@ -15,8 +15,8 @@
 {%- endblock javascript %}
 
 {%- block page_body %}
-<div class="ui container fluid" id="dashboard-user-header-container">
-  <div class="ui container">
+<div class="ui container fluid page-subheader-outer shadowless mb-0 rel-pt-2" id="dashboard-user-header-container">
+  <div class="ui container page-subheader">
     <div class="ui items">
       <div class="item">
 <!--        TODO comment out when avatars ready-->
