+++ invenio_app_rdm/records_ui/templates/semantic-ui/invenio_app_rdm/records/details/citation.html
@@ -7,6 +7,7 @@
     it under the terms of the MIT License; see LICENSE file for more details.
 #}
 
+<div class="ui divider"></div>
 <div id="recordCitation" data-record='{{ record | tojson }}'
   data-styles='{{ config.get("RDM_CITATION_STYLES") | tojson }}'
   data-defaultstyle='{{ config.get("RDM_CITATION_STYLES_DEFAULT") | tojson }}'>
