+++ invenio_app_rdm/records_ui/templates/semantic-ui/invenio_app_rdm/records/details/stats.html
@@ -43,8 +43,8 @@
                   <tbody>
                     <tr class="centered">
                       <td class="left aligned">Views
-                        <span class="ui tooltip-popup text-muted"
-                          data-content="Total views."
+                        <span class="ui text-muted"
+                          data-tooltip="Total views."
                         ><i class="question circle outline icon"></i></span>
                       </td>
                       <td>1,175</td>
@@ -52,8 +52,8 @@
                     </tr>
                     <tr class="centered">
                       <td class="left aligned">Downloads
-                        <span class="ui tooltip-popup text-muted"
-                          data-content="Total downloads."
+                        <span class="ui text-muted"
+                          data-tooltip="Total downloads."
                         ><i class="question circle outline icon"></i></span>
                       </td>
                       <td>52</td>
@@ -61,8 +61,8 @@
                     </tr>
                     <tr class="centered">
                       <td class="left aligned">Data volume
-                        <span class="ui tooltip-popup text-muted"
-                          data-content="Total download volume."
+                        <span class="ui text-muted"
+                          data-tooltip="Total download volume."
                         ><i class="question circle outline icon"></i></span>
                       </td>
                       <td>191.3 MB</td>
@@ -70,8 +70,8 @@
                     </tr>
                     <tr class="centered">
                       <td class="left aligned">Unique views
-                        <span class="ui tooltip-popup text-muted"
-                          data-content="Views in one hour user-sessions are counted only once."
+                        <span class="ui text-muted"
+                          data-tooltip="Views in one hour user-sessions are counted only once."
                         ><i class="question circle outline icon"></i></span>
                       </td>
                       <td>880</td>
@@ -79,8 +79,8 @@
                     </tr>
                     <tr class="centered">
                       <td class="left aligned">Unique downloads
-                        <span class="ui tooltip-popup text-muted"
-                          data-content="Downloads in one hour user-sessions are counted only once."
+                        <span class="ui text-muted"
+                          data-tooltip="Downloads in one hour user-sessions are counted only once."
                         ><i class="question circle outline icon"></i></span>
                       </td>
                       <td>46</td>
