+++ invenio_app_rdm/records_ui/templates/semantic-ui/invenio_app_rdm/records/export.html
@@ -23,7 +23,7 @@
 {% block record_content %}
 <h3>{{ export_format }} Export</h3>
 <div class="top-padded">
-  <pre class="export result">
+  <pre class="export result mb-10 ml-10">
   {{- exported_record -}}
   </pre>
 </div>
