+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/landing_page/RecordCitationField.js
@@ -10,7 +10,13 @@ import axios from "axios";
 import _debounce from "lodash/debounce";
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-import { Header, Placeholder, Grid, Dropdown } from "semantic-ui-react";
+import {
+  Header,
+  Placeholder,
+  Grid,
+  Dropdown,
+  Message,
+} from 'semantic-ui-react';
 import { withCancel } from "../utils";
 import { CopyButton } from "../components/CopyButton";
 import { i18next } from "@translations/invenio_app_rdm/i18next";
@@ -49,9 +55,9 @@ export class RecordCitationField extends Component {
 
   errorMessage = (message) => {
     return (
-      <div className="citation-error-message">
-        <p>{message}</p>
-      </div>
+      <Message negative>
+        {message}
+      </Message>
     );
   };
 
@@ -107,14 +113,9 @@ export class RecordCitationField extends Component {
     });
 
     return (
-      <Grid className="record-citation">
+      <Grid className="record-citation m-0">
         <Grid.Row verticalAlign="middle" className="relaxed">
-          <Grid.Column
-            mobile={8}
-            tablet={8}
-            computer={12}
-            className="no-padding"
-          >
+          <Grid.Column mobile={8} tablet={8} computer={12} className="p-0">
             <h2 id="citation-heading">{i18next.t("Citation")}</h2>
           </Grid.Column>
 
@@ -122,34 +123,34 @@ export class RecordCitationField extends Component {
             mobile={8}
             tablet={8}
             computer={4}
-            className="no-padding"
+            className="p-0"
             textAlign="right"
           >
-            <div className="citation-style-selector">
-              <label id="citation-style-label">{i18next.t("Style")}</label>
-              <Dropdown
-                className="citation-dropdown"
-                aria-labelledby="citation-style-label"
-                defaultValue={defaultStyle}
-                options={citationOptions}
-                selection
-                onChange={_debounce(
-                  (event, data) => this.getCitation(record, data.value),
-                  500
-                )}
-              />
-            </div>
+            <label id="citation-style-label" className="mr-10">
+              {i18next.t("Style")}
+            </label>
+            <Dropdown
+              className="citation-dropdown"
+              aria-labelledby="citation-style-label"
+              defaultValue={defaultStyle}
+              options={citationOptions}
+              selection
+              onChange={_debounce(
+                (event, data) => this.getCitation(record, data.value),
+                500
+              )}
+            />
           </Grid.Column>
         </Grid.Row>
 
         <Grid.Row verticalAlign="bottom">
-          <Grid.Column computer={12} className="no-padding">
+          <Grid.Column computer={12} className="p-0">
             <div id="citation-text">
               {loading ? this.placeholderLoader() : citation}
             </div>
           </Grid.Column>
 
-          <Grid.Column computer={4} className="no-padding" textAlign="right">
+          <Grid.Column computer={4} className="p-0" textAlign="right">
             <CopyButton text={citation} />
           </Grid.Column>
         </Grid.Row>
