+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/landing_page/ShareModal.js
@@ -224,7 +224,7 @@ export const ShareModal = (props) => {
 
         </div>
         <Modal.Description>
-          <p className="share-description">
+          <p className="share-description rel-m-1">
             <Icon name="warning circle" />
             {!!!accessLinkObj
               ? i18next.t(
