+++ invenio_app_rdm/theme/assets/semantic-ui/less/invenio_app_rdm/landing_page/creatibutors.less
@@ -12,6 +12,7 @@
   margin: 0 0 0.5rem 0;
   display: flex;
   flex-wrap: wrap;
+
   &:last-child {
     margin-bottom: 0;
   }
@@ -29,9 +30,11 @@
     color: @mutedTextColor;
     display: flex;
     flex-wrap: wrap;
+
     &:nth-child(n):not(:last-child) {
       margin-right: 1rem;
     }
+
     .identifier-link, .group.icon {
       margin: 0 .1rem 0 .2rem;
     }
