+++ invenio_app_rdm/theme/assets/semantic-ui/less/invenio_app_rdm/landing_page/licenses.less
@@ -10,8 +10,6 @@
 
 section#licenses ul.details-list {
   list-style-type: none;
-  margin: 0;
-  padding: 0;
 
   li {
     margin: .5rem 0;
