+++ invenio_app_rdm/theme/assets/semantic-ui/less/invenio_app_rdm/theme/elements/container.overrides
@@ -2,52 +2,56 @@
          Invenio App RDM Container Overrides
 ***********************************************/
 
-
-.ui.container {
-  &#dashboard-user-header-container {
-    padding-bottom: @defaultPadding;
-    padding-top: @defaultPadding;
-    background-color: @highlightBackgroundColor;
-  }
-  &#dashboard-menu-container {
-    background-color: @highlightBackgroundColor;
-  }
-}
-
-.ui.container.community-header-outer {
-  background-color: @communityHeaderColor;
+.ui.container.page-subheader-outer {
+  background-color: @highlightHeaderColor;
   padding-top: @celledPadding;
   padding-bottom: @celledPadding;
   margin-bottom: @defaultMargin;
 
+  &.with-submenu {
+    padding-bottom: 0;
+    margin-bottom: 0;
+  }
+
   &:not(.compact):not(.with-submenu){
-    box-shadow: 0 0 0 1px rgba(34, 36, 38, 0.22) inset;
+    box-shadow: 0 -@secondaryPointingBorderWidth 0 @borderColor inset;
   }
 
   &.compact {
     height: 5em;
-    .ui.container.community-header {
+    .ui.container.page-subheader {
       height: 3em;
     }
   }
 }
 
-.ui.container.community-header{
-  &.compact {
-    height: 3em;
-  }
+.ui.container.page-subheader {
   align-items: center;
   display: flex;
 
 
-  .community-header-element{
+  &.compact {
+    height: 3em;
+  }
+
+  .page-subheader-element{
     padding-left: 0.5em;
     padding-right: 0.5em;
+
     &:first-of-type {
       padding-left: 0;
     }
     &:last-of-type {
       padding-right: 0;
     }
+  }
+}
+
+
+.ui.container{
+  .frontpage-header{
+    display: flex;
+    align-items: center;
+    height: 20em;
   }
 }
