+++ invenio_app_rdm/theme/assets/semantic-ui/less/invenio_app_rdm/theme/elements/container.variables
@@ -4,3 +4,4 @@
 
 @highlightBackgroundColor: #f8f8f9;
 @celledPadding: 1em;
+@secondaryPointingBorderWidth: 2px;
