+++ invenio_app_rdm/theme/assets/semantic-ui/less/invenio_app_rdm/theme/elements/image.overrides
@@ -2,7 +2,6 @@
          Invenio app RDM Image Overrides
 ***********************************************/
 
-
 .community-logo-header {
   width: 3em;
   height: 3em;
