+++ invenio_app_rdm/theme/assets/semantic-ui/less/invenio_app_rdm/theme/modules/tab.overrides
@@ -1,12 +1,3 @@
 /***********************************************
     Invenio App RDM Tab Overrides
 ***********************************************/
-
-
-.ui.menu.secondary.pointing {
-  .item.active {
-    border-color: @secondaryColor;
-  }
-}
-
-
