+++ invenio_app_rdm/theme/templates/semantic-ui/invenio_app_rdm/header_login.html
@@ -27,7 +27,7 @@
         {# Quick-create ("plus") menu #}
         {%- set plus_menu_items = current_menu.submenu('plus').children %}
         {%- if plus_menu_items %}
-            <div class="rdm-plus-menu rdm-plus-menu-responsive ui dropdown floating">
+            <div class="rdm-plus-menu rdm-plus-menu-responsive ui dropdown floating pr-15">
                 <i class="fitted plus icon inverted"></i>
                 <i class="fitted dropdown icon inverted"></i>
                 <div class="menu">
