+++ invenio_app_rdm/theme/templates/semantic-ui/invenio_app_rdm/javascript.html
@@ -1,11 +1,11 @@
 {# -*- coding: utf-8 -*-
 
   This file is part of Invenio.
-  Copyright (C) 2015-2018 CERN.
+  Copyright (C) 2015-2022 CERN.
 
   Invenio is free software; you can redistribute it and/or modify it
   under the terms of the MIT License; see LICENSE file for more details.
 #}
 {{ webpack['theme.js']}}
-{{ webpack['base-rdm.js']}}
+{{ webpack['base-theme-rdm.js']}}
 {{ webpack['i18n_app.js']}}
